fix: move respondent switch env to the end of the form instead of at the start
